week 2/week2-agent-tools-function-calling/Day 6/tools - approach 2/requirements.txt
openai>=1.9.0
langchain
langchain-openai
langchain-chroma
langchain-community
chromadb
python-dotenv
PyPDF2
tiktoken